requirements.txt
alembic==1.4.2
Flask==1.1.4
Flask-Migrate==2.5.3
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.1
redis==3.5.2
requests==2.23.0
rq==1.8.1
SQLAlchemy==1.3.16
urllib3==1.25.9
Werkzeug==1.0.1
numpy~=1.19.0
gunicorn
amazon-braket-sdk
markupsafe==2.0.1
boto3
botocore